# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Henohenoapi
  description: >-
    The project aims to create a platform for managing exams, their submissions and the reviews related to the submissions
host: henoheno-app.herokuapp.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  Api_key:
    type: apiKey
    in: query
    name: token
security:
  - Api_key: []
paths:
...
  '/users/{user_id}/exams':
    get:
      summary: get the exams of a student
      description: displays the exams created by an user or assigned to him
      tags:
        - Users_resource
      parameters:
        - $ref: '#/parameters/User_id_in_path'
        - $ref: '#/parameters/Selection_users_exams'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exams'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
...
  /groups:
    get:
      summary: groups list
      description: gets the list of all groups
      tags:
        - Groups_resource
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Groups'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: create new group
      description: creates a new group made-up by the members indentified by their id in the payload
      tags:
        - Groups_resource
      parameters:
        - $ref: '#/parameters/Group_in_post'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Group'
        '204':
          $ref: '#/responses/204_no_content'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/groups/{group_id}':
    get:
      summary: get a specific group
      description: access the details of the group specified by the id
      tags:
        - Groups_resource
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Group'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: update an existing group
      description: update an existing group
      tags:
        - Groups_resource
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
        - $ref: '#/parameters/Group_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
    delete:
      summary: delete an existing group
      description: delete an existing group
      tags:
        - Groups_resource
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
      responses:
        '204':
          $ref: '#/responses/204_no_content'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '405':
          $ref: '#/responses/405_method_not_allowed'
        '500':
          $ref: '#/responses/500_internal_server_error'
  /exams:
    get:
      summary: exams list
      description: >-
        displays the exams created by or assigned to the user requesting the
        service
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Selection_exams'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exams'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: create an exam
      description: >-
        creates an exam given the posted informations(tasks and a group should
        already exist)
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Exam_in_post'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Exam'
        '204':
          $ref: '#/responses/204_no_content'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/exams/{id}':
    get:
      summary: access a specific exam
      description: >-
        displays all the informations of an exam object(tasks won't contain the
        solutions)
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: update an existing exam
      description: update an exam by modifying specific fields
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
        - $ref: '#/parameters/Exam_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/exams/{id}/exam_submissions':
    get:
      summary: get the submissions of an exam
      description: displays the exam submission list of a given exam
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_submissions'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
  /exam_submissions:
    get:
      summary: Exam submission list
      description: displays the list of the submissions based on the given criterion
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Selection_exam_submissions'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_submissions'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: Create a new submission
      description: creates new submission based on the posted data
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_in_post'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Exam_submission'
        '204':
          $ref: '#/responses/204_no_content'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/exam_submissions/{exam_submission_id}':
    get:
      summary: access a specific submission
      description: displays the data of a specific submission
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_submission'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: update an existing solution's answers
      description: >-
        As a submitter you might update your submission
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
        - $ref: '#/parameters/Exam_submission_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
    patch:
      summary: assign evaluation
      description: >-
        As the owner of the exam you might want to assign an evaluation to a submission
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
        - $ref: '#/parameters/Exam_submission_in_patch'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/exam_submissions/{exam_submission_id}/exam_peer_reviews':
    get:
      summary: access the reviews of a solution
      description: get all reviews of a specific submission
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_peer_review_for_given_submission'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
  /exam_peer_reviews:
    get:
      summary: reviews list
      description: As a student you can get the list of all your reviews(created or received)
      tags:
        - Exam_peer_reviews_resource
      parameters:
        - $ref: '#/parameters/Selection_reviews'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_peer_reviews'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
    post:
      summary: submit a review
      description: submit a review for a given submission(indentified by its id)
      tags:
        - Exam_peer_reviews_resource
      parameters:
        - $ref: '#/parameters/Exam_peer_review_in_post'
      responses:
        '201':
          description: Created
          schema:
            type: integer
            description: the new exam peer review's id
        '204':
          $ref: '#/responses/204_no_content'
        '400':
          $ref: '#/responses/400_bad_request'
        '401':
          $ref: '#/responses/401_unauthorized'
        '500':
          $ref: '#/responses/500_internal_server_error'
  '/exam_peer_reviews/{exam_peer_review_id}':
    get:
      summary: read a specific review
      description: read a specific review
      tags:
        - Exam_peer_reviews_resource
      parameters:
        - $ref: '#/parameters/Exam_peer_review_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_peer_review'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '500':
          $ref: '#/responses/500_internal_server_error'
    put:
      summary: update a review
      description: update a review
      tags:
        - Exam_peer_reviews_resource
      parameters:
        - $ref: '#/parameters/Exam_peer_review_id_in_path'
        - $ref: '#/parameters/Exam_peer_review_in_put'
      responses:
        '200':
          $ref: '#/responses/200_ok_only'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '409':
          $ref: '#/responses/409_conflict'
        '500':
          $ref: '#/responses/500_internal_server_error'
    delete:
      summary: delete an existing review
      description: delete an existing review
      tags:
        - Exam_peer_reviews_resource
      parameters:
        - $ref: '#/parameters/Exam_peer_review_id_in_path'
      responses:
        '204':
          $ref: '#/responses/204_no_content'
        '401':
          $ref: '#/responses/401_unauthorized'
        '404':
          $ref: '#/responses/404_not_found'
        '405':
          $ref: '#/responses/405_method_not_allowed'
        '500':
          $ref: '#/responses/500_internal_server_error'
#definitions of different objects
definitions:
  Error:
    properties:
      code:
        type: integer
      message:
        type: string
...
  User_in_object:
    type: object
    properties:
      id:
        type: integer
      email:
        type: string
  Exam:
    type: object
    properties:
      id:
        type: integer
      owner:
        $ref: '#/definitions/User_in_object'
      title:
        type: string
      description:
        type: string
      taskset:
        type: array
        items:
          $ref: '#/definitions/Task_in_exam'
      group:
        $ref: '#/definitions/Group'
      final_deadline:
        type: string
        format: date
        description: date format "yyyy-mm-dd"
      review_deadline:
        type: string
        format: date
        description: 'date format "yyyy-mm-dd", review must be at least a day later'
  Exams:
    type: array
    items:
      $ref: '#/definitions/Exam'
...
  Task_in_exam:
    type: object
    properties:
      id:
        type: integer
      description:
        type: string
  Group:
    type: object
    properties:
      id:
        type: integer
      owner:
        $ref: '#/definitions/User_in_object'
      name:
        type: string
      description:
        type: string
      members:
        type: array
        items:
          $ref: '#/definitions/User_in_object'
  Groups:
    type: array
    items:
      $ref: '#/definitions/Group'
  Exam_submission:
    type: object
    properties:
      id:
        type: integer
      ref_exam:
        $ref: '#/definitions/Exam'
      submitter:
        $ref: '#/definitions/User_in_object'
      answers:
        type: array
        items:
          type: string
      status:
        type: string
      evaluation:
        type: string
  Exam_submissions:
    type: array
    items:
      $ref: '#/definitions/Exam_submission'
  Exam_peer_review:
    type: object
    properties:
      id:
        type: integer
      reviewer:
        $ref: '#/definitions/User_in_object'
      submission:
        $ref: '#/definitions/Exam_submission'
      review:
        type: string
  Exam_peer_reviews:
    type: array
    items:
      $ref: '#/definitions/Exam_peer_review'
  Exam_peer_review_for_given_submission:
    type: object
    properties:
      id:
        type: integer
      reviewer:
        $ref: '#/definitions/User_in_object'
      review:
        type: string
...
  Exam_for_post:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      tasks_ids:
        type: array
        description: 'array format "1,2,3,4,5"'
        items:
          type: integer
      group_id:
        type: integer
      final_deadline:
        type: string
        format: date
        description: date format "yyyy-mm-dd"
      review_deadline:
        type: string
        format: date
        description: 'date format "yyyy-mm-dd", review must be at least a day later'
...
  Exam_submission_for_post:
    type: object
    properties:
      ref_exam:
        description: the id of the exam
        type: integer
      answers:
        type: array
        items:
          type: string
      status:
        type: string
  Exam_submission_for_put:
    type: object
    properties:
      answers:
        type: array
        items:
          type: string
      status:
        type: string
  Exam_submission_for_patch:
    type: object
    properties:
      evaluation:
        type: string
  Exam_peer_review_for_post:
    type: object
    properties:
      ref_submission:
        type: integer
      review:
        type: string
  Group_for_post:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      members_id:
        type: array
        items:
          type: integer
  Exam_peer_review_for_put:
    type: object
    properties:
      review:
        type: string
...
parameters:
  Selection_exam_submissions:
    in: query
    name: select
    description: select a particular type of submissions
    required: true
    type: string
    enum:
      - submitted
      - owned
      - toreview
      - reviewed
  Selection_users_exams:
    in: query
    name: select
    description: select a particular type of exams
    required: true
    type: string
    enum:
      - created
      - assigned
  Selection_exams:
    in: query
    name: select
    description: select a particular type of exams
    required: true
    type: string
    enum:
      - created
      - assigned
...
  Selection_reviews:
    in: query
    name: select
    description: select a particular type of reviews
    required: true
    type: string
    enum:
      - created
      - received
...
  User_id_in_path:
    name: user_id
    description: the user's id
    in: path
    type: integer
    required: true
  Exam_in_post:
    name: exam
    description: the object exam
    in: body
    schema:
      $ref: '#/definitions/Exam_for_post'
      required:
        - title
        - description
        - tasks_ids
        - group_id
        - final_deadline
        - review_deadline
  Exam_in_put:
    name: exam
    description: the object exam
    in: body
    schema:
      $ref: '#/definitions/Exam_for_post'
      required:
        - title
        - description
        - tasks_ids
        - group_id
        - final_deadline
        - review_deadline
  Exam_id_in_path:
    name: id
    description: id esame
    in: path
    type: integer
    required: true
...
  Group_in_post:
    name: group
    description: the object group
    in: body
    schema:
      $ref: '#/definitions/Group_for_post'
      required:
        - name
        - description
        - members_id
  Group_in_put:
    name: group
    description: the object group
    in: body
    schema:
      $ref: '#/definitions/Group_for_post'
      required:
        - name
        - description
        - members_id
  Group_id_in_path:
    name: group_id
    description: the id of a group
    in: path
    type: integer
    required: true
  Exam_submission_in_post:
    name: exam_submission
    description: the object exam_submission
    in: body
    schema:
      $ref: '#/definitions/Exam_submission_for_post'
      required:
        - ref_exam
        - answers
  Exam_submission_in_put:
    name: exam_submission
    description: the object exam_submission
    in: body
    schema:
      $ref: '#/definitions/Exam_submission_for_put'
      required:
        - answers
        - status
  Exam_submission_in_patch:
    name: exam_submission
    description: the object exam_submission
    in: body
    schema:
      $ref: '#/definitions/Exam_submission_for_patch'
      required:
        - evaluation
  Exam_submission_id_in_path:
    name: exam_submission_id
    description: exam submission id
    in: path
    type: integer
    required: true
  Exam_peer_review_in_post:
    name: exam_peer_review
    description: the object exam peer review
    in: body
    schema:
      $ref: '#/definitions/Exam_peer_review_for_post'
      required:
        - ref_submission
        - review
  Exam_peer_review_in_put:
    name: exam_peer_review
    description: the object exam peer review
    in: body
    schema:
      $ref: '#/definitions/Exam_peer_review_for_put'
      required:
        - review
  Exam_peer_review_id_in_path:
    name: exam_peer_review_id
    description: exam peer review id
    in: path
    type: integer
    required: true
...
responses:
...
  200_ok_only:
    description: OK
...
  204_no_content:
    description: No Content
  400_bad_request:
    description: Bad Request
    schema:
      $ref: '#/definitions/Error'
  401_unauthorized:
    description: 'Unauthorized, missing or invalid API Key'
    schema:
      $ref: '#/definitions/Error'
...
  404_not_found:
    description: Not Found
    schema:
      $ref: '#/definitions/Error'
  405_method_not_allowed:
    description: Method Not Allowed
    schema:
      $ref: '#/definitions/Error'
  409_conflict:
    description: Conflict
    schema:
      $ref: '#/definitions/Error'
  500_internal_server_error:
    description: Internal Server Error
    schema:
      $ref: '#/definitions/Error'